requirements.txt
# AI Integration Tool - Dependencies
requests>=2.28.0
pyyaml>=6.0
tenacity>=8.2.0
tiktoken>=0.5.0
httpx>=0.25.0
